+++ src/scene_cluster_lod.cpp
@@ -287,16 +287,20 @@ int Scene::clodGroupMeshoptimizer(void* output_context, clodGroup group, const c
 
 void Scene::buildGeometryClusterLodMeshoptimizer(const ProcessingInfo& processingInfo, GeometryStorage& geometry)
 {
-  clodConfig clodInfo     = m_config.meshoptPreferRayTracing ? clodDefaultConfigRT(m_config.clusterTriangles) :
-                                                               clodDefaultConfig(m_config.clusterTriangles);
-  clodInfo.max_vertices   = m_config.clusterVertices;
-  clodInfo.partition_size = m_config.clusterGroupSize;
+  clodConfig clodInfo = m_config.meshoptPreferRayTracing ? clodDefaultConfigRT(m_config.clusterTriangles) :
+                                                           clodDefaultConfig(m_config.clusterTriangles);
+
+  clodInfo.max_vertices              = m_config.clusterVertices;
+  clodInfo.partition_size            = m_config.clusterGroupSize;
+  clodInfo.partition_spatial_average = m_config.clusterGroupSize / 2;
 
   // account for meshopt_partitionClusters's using a target value with a higher worst case
   while((clodInfo.partition_size + clodInfo.partition_size / 3) > m_config.clusterGroupSize)
   {
     clodInfo.partition_size--;
   }
+
+  clodInfo.partition_spatial_average = std::min(clodInfo.partition_spatial_average, clodInfo.partition_size - 1);
 
   // These control the error propagation across lod levels to
   // account for simplifying an already simplified mesh.
@@ -344,8 +348,17 @@ void Scene::buildGeometryClusterLodMeshoptimizer(const ProcessingInfo& processin
     uint32_t groupCount =
         uint32_t(geometry.lodMesh.groupClusterRanges.size()) - geometry.lodMesh.lodLevelGroupRanges.back().offset;
 
-    geometry.lodMesh.lodLevelGroupRanges.back().count = groupCount;
-    geometry.lodLevels.back().groupCount              = groupCount;
+    nvcluster_Range& lastLodLevelGroupRange = geometry.lodMesh.lodLevelGroupRanges.back();
+
+    lastLodLevelGroupRange.count         = groupCount;
+    geometry.lodLevels.back().groupCount = groupCount;
+
+    if(lastLodLevelGroupRange.count != 1 || geometry.lodMesh.groupClusterRanges[lastLodLevelGroupRange.offset].count != 1)
+    {
+      assert(0);
+      LOGE("clodBuild failed: last lod level has more than one cluster\n");
+      std::exit(-1);
+    }
   }
 
   // vectors are resized at end of lod processing,